First pass interpretation of rspec summary
Selected what I thought was the first section.
Took a screenshot of it. It was intimidating.
Copied and pasted the most intimidating section, and higlighted all
of the intimidating words.  It was slightly less intimidating.

Context: am fascinated by this, and if I don't stop with this commit
message I am going to be late for my group meeting, which would be a
Bad Thing.  It would show, as DBC puts it, a lack of integrity.