Added section on body of rspec error message
So far, in this section, I have inspected the description and
what I assume is the source code of the text.

Context: Drinking yesterday's tea and slouching. Kevin and Jacky are
still walking to the dounut vautl.  There is blessed silence and only
one other person working in the space. My section of the room has the
lights off and it is snowing.  I need to sit up straight, but man I
feel productive right now.