The name attribute in the html is the key in the params hash
The params function is a gumball machine that, I theorize, takes in a quarter that is in fact a raw html form.  Inside the params function, I assume there are little elves that take apart the form. The form is a box with lots of boxes inside boxes, like some mysterious woodworker had a lot of time on his hands.  Several of the boxes are labeled input. Inside those input boxes are boxes labeled name.  The elves take out the boxes labeled name and look inside.  Inside is a bunch string with a bunch of letter beads strung on it, like kid's frendship bracelets with their names spelled out. The elves take out another set of safe boxes.  Each of those save boxes has a lock. On that lock is a key.  The elves copy the text from the friendship bracelet onto the key. They look in the mysterioys wooden box again. Right next to where they got the boxes labeled 'name', they find boxes labeled value.  They take the box labeled value, open it up, take out the friendship bracelet inside, and stuff it inside the safety deposit box. When a safety deposit box is full, they shove the box into a bag that smells like hash-browns.  They stick the bag inside a gumball and send it out the door.