Input values to array of values
The elves look into the input box and see a label square brackets
at the end of the label. For example

 name = "stingray[]"

The elves know that when they see a name with square brackets, they are
supposed to copy the name to a single safety deposite box.  They then
take the friendship bracets out of every value box that is inside an
input box with the name stingray.  They stick every friendship bracelet
inside a pillbox. Each bracelet gets it's own compartment in the pillbox.
Then the elves stick the pillbox in teh safety deposite box, stick the
box in a bag, wrap the box in a gumball, and shove it out the door to
the waiting gumball buyer.