You can nest arrays inside the hashes
The elves have recieved a new wooden puzzle box and are getting more and more experienced at handling strange cases. When they see two sets of square braces on a name box, like

name="sharkbait[nemo][]"

they take out two sets of safety deposite boxes and a pillbox. They write sharkbait on the key to the biggest safety deposit box.  They write nemo on the key to the smaller safety box.  They take out a valuable thing from the values wooden box.  I imagine it's a set of earrings.  They stick the earrings in the pill box.  They stick the pillbox in teh small safty deposit box. The stick the small safety deposit box in the large safety deposit box. They stick the box in a bag.  They stick the bag in a gumball and shove it out the door.