Putting a hash inside an array inside a hash
The elves have recieved their final wooden box.  They open the boxes within the boxes, and they see a name that looks like this:

  name=sharkbait[][nemo]

They get out a safety deposit box, a large pillbox, and another safety deposit box that fits inside one cell of the pillbox.  They write sharkbait on the key to the large safety deposit box.  They write 1 on one of the cells of the pillbox.  They write nemon on the key to the small safety deposit box. They get a set of earrings from the values box and stick them in the small safety deposit box, stick the small safety deposit box into the pillbox named 1, and stick the pillbox in the large safety deposite box, stick the large safty deposite box in a bag, stick teh bag in a gumball, and roll the gumball out the door.